Get braille files for each item from SB FTP server
Smarter Balanced supports several braille types, but not all items have braille
files for all of the braille types. The SB FTP server has the braille files. We
are checking which braille files exsist on the server for each item. This adds ~20 minutes to app start so we will probably take
a different route with this. I want to commit this code in case we do not.